refactor(ui): split useCreateDeploy.ts god object into modules
