+++ scripts/.god-object-baseline.txt
@@ -13,5 +13,4 @@
 701 ui/src/lib/stores/chatSessions.ts
 682 ui/src/components/shared/InfrastructureModal.tsx
 653 ai-agent-instance-blueprint-lib/src/auto_provision.rs
-627 ui/src/lib/hooks/useCreateDeploy.ts
 606 sandbox-runtime/src/util.rs
